Update reputation.lua for TWW
This one has more extensive changes, as the warband addition and reputation panel rework required it. Menu had to be rewritten.

Some minor improvements are made, but I would like to improve some things in subsequent releases. For example the way the reputation list is generated.

Also added a "Auto switch bar to last gained reputation" option, which automatically switches the Reputation bar to the reputation you last gained rep with.

This should work for TWW though, and has no noticeable issues or performance impacts.